pyproject.toml
[tool.poetry]
name = "gcp-utils"
version = "1.0.0"
description = "Utilities for Google Cloud Platform"
authors = ["v-crn <var.cyrano@gmail.com>"]

[virtualenvs]
in-project = false

[tool.poetry.dependencies]
python = ">=3.9"
cloudpickle = "^2.2.1"
gcsfs = "*"
google-cloud-bigquery = "*"
google-cloud-secret-manager = "^2.16.3"
google-cloud-storage = "*"
numpy = "*"
pandas = "*"
tqdm = "*"

[tool.poetry.group.dev.dependencies]
autoflake = "*"
black = "*"
click = "*"
flake8 = "*"
ipywidgets = "*"
isort = "*"
jupyter-kite = "*"
jupyterlab = "*"
jupyterlab-code-formatter = "*"
jupyterlab-flake8 = "*"
jupyterlab-git = "*"
jupyterlab-lsp = "*"
jupyterlab-widgets = "*"
jupytext = "*"
lckr-jupyterlab-variableinspector = "*"
mypy = "*"
pyproject-flake8 = "*"
types-python-dateutil = "*"

[tool.poetry.group.test.dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.flake8]
max-line-length = 88
ignore = "E203,W503,W504"
max-complexity = 20

[tool.isort]
profile = "black"

[tool.mypy]
strict_optional = "True"
warn_redundant_casts = "True"
warn_unused_ignores = "True"
disallow_any_generics = "True"
check_untyped_defs = "True"
disallow_untyped_defs = "True"
ignore_missing_imports = "True"

[[tool.mypy.overrides]]
module = ["matplotlib.*", "tqdm.*", "pandas.*", "scipy"]
ignore_missing_imports = "True"
